// src/app/models/enums/ViewUrls.ts
export enum ViewUrls {
  ABOUT_US = "about-us",
  ACCOUNT_SETTINGS = "account-settings",
  ADMIN_BESTELLUNG_DETAILS = "admin-bestellung-details",
  ADMIN_BESTELLUNGEN_OVERVIEW = "admin-bestellungen-overview",
  ADMIN_PRODUCT_DETAILS = "admin-product-details",
  ADMIN_PRODUCTS_OVERVIEW = "admin-products-overview",
  BESTELLUNG_DETAILS = "bestellung-details",
  BESTELLUNGEN_OVERVIEW = "bestellungen-overview",
  CHECKOUT = "checkout",
  FAVORITEN_LISTE = "favoriten-liste",
  PAYMENT_APPROVAL = "payment-approval",
  PRODUKT_DETAILS = "produkt-details",
  PRODUKTE_OVERVIEW = "produkt-overview",
  WARENKORB = "warenkorb"
}